Move pause para menu.h
